requirements.txt

altgraph==0.17.2
future==0.18.2
macholib==1.15.2
six==1.15.0
annotated-types==0.7.0
anyio==4.1.0
appnope==0.1.4
APScheduler==3.6.3
asttokens==3.0.0
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==4.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.8
decorator==5.1.1
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.0
executing==2.1.0
Flask==3.1.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-play-scraper==1.2.7
gspread==6.1.4
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.28.1
idna==3.6
importlib_metadata==8.5.0
instaloader==4.14
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
nest-asyncio==1.6.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.82.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pd==0.0.4
pillow==11.2.1
platformdirs==4.3.6
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
pyTelegramBotAPI==4.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2024.2
pyzmq==26.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.27.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
tornado==6.1
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
urllib3==1.26.16
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
opencv-python-headless==4.9.0.80
matplotlib==3.8.4
PyDrive==1.3.1
google-api-python-client==2.126.0
oauth2client==4.1.3
